feat: Rename to service
